shell/db - minor refactor
