--- rust/shell/src/db/models/mod.rs
@@ -1,6 +1,0 @@
-mod post;
-mod user;
-
-// flat re-export
-pub use post::*;
-// pub use user::*;
